* Log datatype should be tinyint (mysql install/upgrade sql) [Bug 3910]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@9539 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ other/install_2-0_mysql.sql
@@ -834,7 +834,7 @@ CREATE TABLE {$db_prefix}log_digest (
   id_topic mediumint(8) unsigned NOT NULL,
   id_msg int(10) unsigned NOT NULL,
   note_type varchar(10) NOT NULL default 'post',
-  daily smallint(3) unsigned NOT NULL default '0',
+  daily tinyint(3) unsigned NOT NULL default '0',
   exclude mediumint(8) unsigned NOT NULL default '0'
 ) ENGINE=MyISAM;
